# internal/swagger/swagger.yaml
definitions:
  errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.Dictionary:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
  schema.DictionaryForm:
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: A test API service based on golang.
  title: testapp
  version: v1.0.0
paths:
...
  /api/v1/sys/dictionaries:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        required: true
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Dictionary'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query dictionary list
      tags:
      - DictionaryAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.DictionaryForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Dictionary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create dictionary record
      tags:
      - DictionaryAPI
  /api/v1/sys/dictionaries/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete dictionary record by ID
      tags:
      - DictionaryAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Dictionary'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get dictionary record by ID
      tags:
      - DictionaryAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.DictionaryForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update dictionary record by ID
      tags:
      - DictionaryAPI
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"